+++ test_runner/src/test/kotlin/ftl/reports/utils/ReportManagerTest.kt
@@ -8,6 +8,7 @@ import ftl.reports.FullJUnitReport
 import ftl.reports.JUnitReport
 import ftl.reports.MatrixResultsReport
 import ftl.reports.util.ReportManager
+import ftl.reports.util.getMatrixPath
 import ftl.reports.xml.model.JUnitTestCase
 import ftl.reports.xml.model.JUnitTestResult
 import ftl.reports.xml.model.JUnitTestSuite
@@ -21,11 +22,13 @@ import io.mockk.mockkObject
 import io.mockk.unmockkAll
 import io.mockk.verify
 import org.junit.After
+import org.junit.Assert.assertEquals
 import org.junit.Rule
 import org.junit.Test
 import org.junit.contrib.java.lang.system.SystemErrRule
 import org.junit.contrib.java.lang.system.SystemOutRule
 import org.junit.runner.RunWith
+import java.io.File
 
 @RunWith(FlankTestRunner::class)
 class ReportManagerTest {
@@ -195,5 +198,14 @@ class ReportManagerTest {
 
         assertThat(ReportManager.getDeviceString(""))
             .isEqualTo("")
+    }
+
+    @Test
+    fun `should get weblink from legacy path and ios path`() {
+        val legacyPath = File("results/2020-08-06_12-08-55.641213_jGpY/matrix_0/NexusLowRes-28-en-portrait/test_result_1.xml")
+        val iosPath = File("results/test_dir/shard_0/iphone8-12.0-en-portrait/test_result_0.xml")
+        val objectName = "object_name"
+        assertEquals("object_name/2020-08-06_12-08-55.641213_jGpY/matrix_0", legacyPath.getMatrixPath(objectName))
+        assertEquals("object_name/test_dir/shard_0", iosPath.getMatrixPath(objectName))
     }
 }